+++ spec/lib/deterministic/result/success_spec.rb
@@ -18,8 +18,6 @@ describe Deterministic::Success do
 
   specify { expect(subject).to be_an_instance_of described_class }
   specify { expect(subject).to eq(described_class.new(1)) }
-  specify { expect(subject << Success(2)).to eq(Success(2)) }
-  specify { expect(subject << Failure(2)).to eq(Failure(2)) }
   specify { expect(subject.fmap { |v| v + 1} ).to eq Success(2) }
 
   specify { expect(subject.or(Success(2))).to eq Success(1)}