+++ spec/lib/deterministic/option_spec.rb
@@ -83,7 +83,7 @@ describe Deterministic::Option::None do
   specify { expect(described_class.new).to eq None }
   specify { expect(described_class.new.some?).to be_falsey }
   specify { expect(described_class.new.none?).to be_truthy }
-  # specify { expect { described_class.new.value }.to raise_error RuntimeError }
+  specify { expect { described_class.new.value }.to raise_error NoMethodError }
 
   specify { expect(described_class.new.fmap { |n| n + 1}).to eq None }
   specify { expect(described_class.new.map { |n| nil }).to eq None }